added type:moudle to package.json to enable dealing with moudles. created logic.js which will store logic of project
