+++ package.json
@@ -3,5 +3,6 @@
   "version": "1.0.0",
   "scripts": {
     "test": "node common.test.mjs"
-  }
+  },
+  "type": "module"
 }
